requirements.txt
fastapi
uvicorn
redis 
pydantic 
httpx 
python-dotenv 
groq